+++ src/main/kotlin/dev/aaronhowser/mods/genetics_resequenced/enchantment/ModEnchantments.kt
@@ -12,9 +12,15 @@ object ModEnchantments {
 
 	val DELICATE_TOUCH: ResourceKey<Enchantment> =
 		ResourceKey.create(Registries.ENCHANTMENT, GeneticsResequenced.modResource("delicate_touch"))
+	val SURGICAL_PRECISION: ResourceKey<Enchantment> =
+		ResourceKey.create(Registries.ENCHANTMENT, GeneticsResequenced.modResource("surgical_precision"))
 
 	fun getDelicateTouchHolder(entity: Entity): Holder<Enchantment> {
 		return OtherUtil.getEnchantHolder(entity, DELICATE_TOUCH)
+	}
+
+	fun getSurgicalPrecisionHolder(entity: Entity): Holder<Enchantment> {
+		return OtherUtil.getEnchantHolder(entity, SURGICAL_PRECISION)
 	}
 
 }